+++ skills/mechanics-sketches/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.3.0] - 2026-05-04
+
+### Added
+- `pressure` element: n radial arrows around a central annotation (Drucksymbol)
+- `abs_length` parameter for `fixed_support`: absolute wall length in scene units, scale-independent
+
 ## [1.2.0] - 2026-04-13
 
 ### Added